add minimum start listener
