+++ src/main/java/io/shiftleft/bctrace/asm/Transformer.java
@@ -30,6 +30,7 @@ import java.lang.instrument.IllegalClassFormatException;
 import java.security.ProtectionDomain;
 import java.util.List;
 import io.shiftleft.bctrace.asm.helper.CatchHelper;
+import io.shiftleft.bctrace.asm.helper.MinStartHelper;
 import io.shiftleft.bctrace.asm.helper.ReturnHelper;
 import io.shiftleft.bctrace.asm.helper.StartArgumentsHelper;
 import io.shiftleft.bctrace.asm.helper.StartHelper;
@@ -38,12 +39,13 @@ import io.shiftleft.bctrace.asm.utils.ASMUtils;
 import io.shiftleft.bctrace.runtime.Callback;
 import io.shiftleft.bctrace.runtime.MethodRegistry;
 import io.shiftleft.bctrace.spi.Hook;
-import io.shiftleft.bctrace.spi.listener.BeforeThrownListener;
-import io.shiftleft.bctrace.spi.listener.FinishReturnListener;
-import io.shiftleft.bctrace.spi.listener.FinishThrowableListener;
+import io.shiftleft.bctrace.spi.listener.info.BeforeThrownListener;
+import io.shiftleft.bctrace.spi.listener.info.FinishReturnListener;
+import io.shiftleft.bctrace.spi.listener.info.FinishThrowableListener;
 import io.shiftleft.bctrace.spi.listener.Listener;
-import io.shiftleft.bctrace.spi.listener.StartArgumentsListener;
-import io.shiftleft.bctrace.spi.listener.StartListener;
+import io.shiftleft.bctrace.spi.listener.info.StartArgumentsListener;
+import io.shiftleft.bctrace.spi.listener.info.StartListener;
+import io.shiftleft.bctrace.spi.listener.min.MinStartListener;
 import java.util.ArrayList;
 import org.objectweb.asm.ClassReader;
 import org.objectweb.asm.ClassWriter;
@@ -162,9 +164,10 @@ public class Transformer implements ClassFileTransformer {
   }
 
   private void modifyMethod(ClassNode cn, MethodNode mn, ArrayList<Integer> hooksToUse) {
-    
+
     int methodId = MethodRegistry.getInstance().getMethodId(cn.name, mn.name, mn.desc);
-    
+
+    ArrayList<Integer> minStartListenerHooks = getListenerHooks(hooksToUse, MinStartListener.class);
     ArrayList<Integer> startListenerHooks = getListenerHooks(hooksToUse, StartListener.class);
     ArrayList<Integer> startArgumentsListenerHooks = getListenerHooks(hooksToUse, StartArgumentsListener.class);
     ArrayList<Integer> finishReturnListenerHooks = getListenerHooks(hooksToUse, FinishReturnListener.class);
@@ -172,6 +175,7 @@ public class Transformer implements ClassFileTransformer {
     ArrayList<Integer> beforeThrownListenerHooks = getListenerHooks(hooksToUse, BeforeThrownListener.class);
 
     LabelNode startNode = CatchHelper.insertStartNode(mn, finishThrowableListenerHooks);
+    MinStartHelper.addTraceStart(methodId, cn, mn, minStartListenerHooks);
     StartHelper.addTraceStart(methodId, cn, mn, startListenerHooks);
     StartArgumentsHelper.addTraceStart(methodId, cn, mn, startArgumentsListenerHooks);
     ReturnHelper.addTraceReturn(methodId, mn, finishReturnListenerHooks);
