+++ src/main/java/io/shiftleft/bctrace/spi/listener/info/BeforeThrownListener.java
@@ -22,24 +22,24 @@
  * CONVEY OR IMPLY ANY RIGHTS TO REPRODUCE, DISCLOSE OR DISTRIBUTE ITS
  * CONTENTS, OR TO MANUFACTURE, USE, OR SELL ANYTHING THAT IT MAY DESCRIBE, IN WHOLE OR IN PART.
  */
-package io.shiftleft.bctrace.spi.listener;
+package io.shiftleft.bctrace.spi.listener.info;
+
+import io.shiftleft.bctrace.spi.listener.Listener;
 
 /**
  *
  * @author Ignacio del Valle Alles idelvall@shiftleft.io
  */
-public interface FinishThrowableListener extends Listener {
+public interface BeforeThrownListener extends Listener {
 
   /**
-   * Invoked by instrumented methods just before rising a throwable to the
-   * caller (if multiple plugins are registered, listener notification is
-   * performed according to their respective plugin <b>reverse</b>
-   * registration order).
+   * Invoked by instrumented methods just before the actual method throws a
+   * throwable.
    *
    * @param methodId method id (as defined by MethodRegistry)
    * @param instance instance where the method belongs. Null if the method is
-   * static
-   * @param th thowable to be raised by the instrumented method
+   * @param th throwable to be thrown
    */
-  public void onFinishedThrowable(int methodId, Object instance, Throwable th);
+  public void onBeforeThrown(int methodId, Object instance, Throwable th);
+
 }
