+++ src/main/java/io/shiftleft/bctrace/spi/listener/info/FinishReturnListener.java
@@ -22,23 +22,28 @@
  * CONVEY OR IMPLY ANY RIGHTS TO REPRODUCE, DISCLOSE OR DISTRIBUTE ITS
  * CONTENTS, OR TO MANUFACTURE, USE, OR SELL ANYTHING THAT IT MAY DESCRIBE, IN WHOLE OR IN PART.
  */
-package io.shiftleft.bctrace.spi.listener;
+package io.shiftleft.bctrace.spi.listener.info;
+
+import io.shiftleft.bctrace.spi.listener.Listener;
 
 /**
+ *
  * @author Ignacio del Valle Alles idelvall@shiftleft.io
  */
-public interface StartArgumentsListener extends Listener {
+public interface FinishReturnListener extends Listener {
 
-/**
-   * Invoked by instrumented methods before any of its original instructions (if
-   * multiple plugins are registered, listener notification is performed
-   * according to their respective plugin registration order).
-   *
+  /**
+   * Invoked by instrumented methods just before return (if multiple plugins are
+   * registered, listener notification is performed according to their
+   * respective plugin <b>reverse</b> registration order).
+   * 
    * @param methodId method id (as defined by MethodRegistry)
    * @param instance instance where the method belongs. Null if the method is
-   * static
-   * @param args arguments passed to the method. 
-   * returns false;
+   * @param ret Object being returned by the method. Wrapper type if the
+   * original return type is primitive. <code>null</code> if the method return
+   * type is <code>void</code>
    */
-  public void onStart(int methodId, Object instance, Object[] args);
+  
+  public void onFinishedReturn(int methodId, Object instance, Object ret);
+
 }
