+++ src/main/java/io/shiftleft/bctrace/spi/listener/info/FinishThrowableListener.java
@@ -22,26 +22,26 @@
  * CONVEY OR IMPLY ANY RIGHTS TO REPRODUCE, DISCLOSE OR DISTRIBUTE ITS
  * CONTENTS, OR TO MANUFACTURE, USE, OR SELL ANYTHING THAT IT MAY DESCRIBE, IN WHOLE OR IN PART.
  */
-package io.shiftleft.bctrace.spi.listener;
+package io.shiftleft.bctrace.spi.listener.info;
+
+import io.shiftleft.bctrace.spi.listener.Listener;
 
 /**
  *
  * @author Ignacio del Valle Alles idelvall@shiftleft.io
  */
-public interface FinishReturnListener extends Listener {
+public interface FinishThrowableListener extends Listener {
 
   /**
-   * Invoked by instrumented methods just before return (if multiple plugins are
-   * registered, listener notification is performed according to their
-   * respective plugin <b>reverse</b> registration order).
-   * 
+   * Invoked by instrumented methods just before rising a throwable to the
+   * caller (if multiple plugins are registered, listener notification is
+   * performed according to their respective plugin <b>reverse</b>
+   * registration order).
+   *
    * @param methodId method id (as defined by MethodRegistry)
    * @param instance instance where the method belongs. Null if the method is
-   * @param ret Object being returned by the method. Wrapper type if the
-   * original return type is primitive. <code>null</code> if the method return
-   * type is <code>void</code>
+   * static
+   * @param th thowable to be raised by the instrumented method
    */
-  
-  public void onFinishedReturn(int methodId, Object instance, Object ret);
-
+  public void onFinishedThrowable(int methodId, Object instance, Throwable th);
 }
